feat: add useDisconnect hook with localStorage cleanup
- Create centralized useDisconnect hook in ui.ts that handles wallet disconnection
- Clear all bridge-related localStorage values on disconnect
- Refactor WidgetHeader to use the new useDisconnect hook
- Export disconnect function from useInterwovenKit for external usage
