+++ packages/interwovenkit-react/src/data/ui.ts
@@ -1,6 +1,8 @@
+import { useDisconnect as useDisconnectWagmi } from "wagmi"
 import { atom, useAtom } from "jotai"
-import { useReset } from "@/lib/router"
+import { useNavigate, useReset } from "@/lib/router"
 import { useAnalyticsTrack } from "@/data/analytics"
+import { LocalStorageKey } from "./constants"
 
 const isDrawerOpenAtom = atom<boolean>(false)
 
@@ -23,4 +25,24 @@ export function useDrawer() {
   }
 
   return { isDrawerOpen: isOpen, openDrawer: open, closeDrawer: close }
+}
+
+export function useDisconnect() {
+  const navigate = useNavigate()
+  const { closeDrawer } = useDrawer()
+  const { disconnect } = useDisconnectWagmi()
+
+  return () => {
+    navigate("/blank")
+    closeDrawer()
+    disconnect()
+
+    // Clear bridge form values on disconnect
+    localStorage.removeItem(LocalStorageKey.BRIDGE_SRC_CHAIN_ID)
+    localStorage.removeItem(LocalStorageKey.BRIDGE_SRC_DENOM)
+    localStorage.removeItem(LocalStorageKey.BRIDGE_DST_CHAIN_ID)
+    localStorage.removeItem(LocalStorageKey.BRIDGE_DST_DENOM)
+    localStorage.removeItem(LocalStorageKey.BRIDGE_QUANTITY)
+    localStorage.removeItem(LocalStorageKey.BRIDGE_SLIPPAGE_PERCENT)
+  }
 }
