+++ packages/interwovenkit-react/src/public/app/WidgetHeader.tsx
@@ -1,13 +1,11 @@
 import clsx from "clsx"
-import { useAccount, useDisconnect } from "wagmi"
+import { useAccount } from "wagmi"
 import { useState, useRef, useEffect } from "react"
 import { useSpring, animated } from "@react-spring/web"
 import { IconCopy, IconQrCode, IconSignOut } from "@initia/icons-react"
 import { truncate } from "@initia/utils"
-import { useNavigate } from "@/lib/router"
 import { useInterwovenKit } from "@/public/data/hooks"
-import { useDrawer } from "@/data/ui"
-import { LocalStorageKey } from "@/data/constants"
+import { useDisconnect } from "@/data/ui"
 import { useDefaultChain } from "@/data/chains"
 import CopyButton from "@/components/CopyButton"
 import Image from "@/components/Image"
@@ -16,12 +14,10 @@ import AddressQr from "./AddressQr"
 import styles from "./WidgetHeader.module.css"
 
 const WidgetHeader = () => {
-  const navigate = useNavigate()
   const deafultChain = useDefaultChain()
   const { connector } = useAccount()
-  const { disconnect } = useDisconnect()
+  const disconnect = useDisconnect()
   const { address, username } = useInterwovenKit()
-  const { closeDrawer } = useDrawer()
   const { openModal } = useModal()
   const name = username ?? address
 
@@ -37,17 +33,7 @@ const WidgetHeader = () => {
     if (!isExpanded) {
       setIsExpanded(true)
     } else {
-      navigate("/blank")
-      closeDrawer()
       disconnect()
-
-      // Clear bridge form values on disconnect
-      localStorage.removeItem(LocalStorageKey.BRIDGE_SRC_CHAIN_ID)
-      localStorage.removeItem(LocalStorageKey.BRIDGE_SRC_DENOM)
-      localStorage.removeItem(LocalStorageKey.BRIDGE_DST_CHAIN_ID)
-      localStorage.removeItem(LocalStorageKey.BRIDGE_DST_DENOM)
-      localStorage.removeItem(LocalStorageKey.BRIDGE_QUANTITY)
-      localStorage.removeItem(LocalStorageKey.BRIDGE_SLIPPAGE_PERCENT)
     }
   }
